Big changes
Needed to be able to open a native select for mobile, so now there’s a 
hidden select element always present which the component tries to open
on click/tap if the device is a touchscreen.

Since we have that select now, might as well use it to size itself, get
the options, get which item is selected, and the values of each option.

Also looks a lot more like a native os x select element