BREAKING Now uses the dropdown to size instead of select
Does this by hoisting the drop downs to a fixed div on the body
and then positions them when open.

Has the benefit of being sized correctly when initially display:none
and also prevents the drop down from being clipped in a scroll container

Downside is that it complicated styling the dropdown if you need a mix of
different styles for each dropdown. As a workaround, classes on the <select>
are mirrored to the dropdown.

Not an issue if you style all dropdowns similarly on the page.